Dhewm3 + RBDOOM3BFG Rebases
Doesn't include the render

+++ neo/framework/Compressor.cpp
@@ -1727,7 +1727,7 @@ idCompressor_Arithmetic::GetCurrentCount
 */
 int idCompressor_Arithmetic::GetCurrentCount()
 {
-	return ( unsigned int )( ( ( ( ( long ) code - low ) + 1 ) * scale - 1 ) / ( ( ( long ) high - low ) + 1 ) );
+	return ( unsigned int )( ( ( ( ( int ) code - low ) + 1 ) * scale - 1 ) / ( ( ( int ) high - low ) + 1 ) );
 }
 
 /*
@@ -1806,9 +1806,9 @@ idCompressor_Arithmetic::RemoveSymbolFromStream
 */
 void idCompressor_Arithmetic::RemoveSymbolFromStream( acSymbol_t* symbol )
 {
-	long range;
+	int range;
 
-	range	= ( long )( high - low ) + 1;
+	range	= ( int )( high - low ) + 1;
 	high	= low + ( unsigned short )( ( range * symbol->high ) / scale - 1 );
 	low		= low + ( unsigned short )( ( range * symbol->low ) / scale );
